Compute polynomial hash of stack trace instead of plain sum. Also, make hash 64-bit on 64-bit platform.

+++ HeapyInject/HeapProfiler.h
@@ -6,7 +6,7 @@
 #include <mutex>
 
 const int backtraceSize = 64;
-typedef unsigned long StackHash;
+typedef size_t StackHash;
 
 struct StackTrace{
 	void *backtrace[backtraceSize];
